requirements.txt
Django==4.2.2
django-ckeditor==6.4.2
django-modeltranslation
gunicorn==20.1.0
Pillow==9.2.0
psycopg2-binary==2.9.3
python-decouple==3.6
